Add error message when delete product doesn't find a product

+++ src/module/product/service/productService.js
@@ -22,6 +22,10 @@ class ProductService {
   }
 
   async delete(product) {
+    if (!(product instanceof Product)) {
+      throw new ProductNotDefinedError();
+    }
+    
     return this.productRepository.delete(product);
   }
 }
